--- DenseNet100_TinyMS/README_CN.md
@@ -1,3 +1,0 @@
-基于GPU版的TinyMS，开始训练：python train.py
-cifar10数据集放置在train.py目录下
-cifar10_path = 'cifar10/cifar-10-batches-bin'
