+++ densenet_BC_100.md
@@ -1,0 +1,9 @@
+重新build之后,将test/st/densenetBC_100.py放置在tinyms同级目录下面</br>
+ ```
+python densenetBC_100.py \
+--num_classes 10 \
+--epoch_size 1 \
+--device_target GPU \
+--dataset_path {cifar10-path}\
+--batch_size 128
+ ```
