+++ tinyms/model/__init__.py
@@ -22,14 +22,16 @@ from .resnet50 import resnet50, ResNet
 from .mobilenetv2 import mobilenetv2, mobilenetv2_infer, MobileNetV2
 from .ssd300 import ssd300_mobilenetv2, ssd300_infer, SSD300
 from .cycle_gan.cycle_gan import cycle_gan, cycle_gan_infer
-
+from .densenetBC_100 import densenetBC_100
+from .alexnet import AlexNet
 __all__ = [
     'Model',
     'lenet5', 'LeNet',
     'resnet50', 'ResNet',
     'mobilenetv2', 'mobilenetv2_infer', 'MobileNetV2',
     'ssd300_mobilenetv2', 'ssd300_infer', 'SSD300',
-    'cycle_gan', 'cycle_gan_infer',
+    'cycle_gan', 'cycle_gan_infer','densenet_BC_100'
+    'AlexNet'
 ]
